# Conflicts:
#	engine/gameobject/src/gameobject/gameobject.cpp
#	engine/gameobject/src/gameobject/gameobject_props.cpp
#	engine/gameobject/src/gameobject/gameobject_props.h
#	engine/gameobject/src/gameobject/res_collection.cpp
#	engine/gameobject/src/gameobject/res_prototype.cpp
#	engine/gamesys/src/gamesys/components/comp_label.cpp
#	engine/gamesys/src/gamesys/components/comp_model.cpp
#	engine/gamesys/src/gamesys/components/comp_model.h
#	engine/gamesys/src/gamesys/components/comp_private.cpp
#	engine/gamesys/src/gamesys/components/comp_sprite.cpp
#	engine/gamesys/src/gamesys/components/comp_tilegrid.cpp
#	engine/gamesys/src/gamesys/gamesys_private.h
#	engine/gamesys/src/gamesys/scripts/script_label.cpp
#	engine/gamesys/src/gamesys/scripts/script_model.cpp
#	engine/gamesys/src/gamesys/scripts/script_sprite.cpp
#	engine/resource/src/resource.h

+++ engine/render/src/render/render.h
@@ -296,7 +296,7 @@ namespace dmRender
     HMaterial                       NewMaterial(dmRender::HRenderContext render_context, dmGraphics::HVertexProgram vertex_program, dmGraphics::HFragmentProgram fragment_program);
     void                            DeleteMaterial(dmRender::HRenderContext render_context, HMaterial material);
     void                            ApplyMaterialConstants(dmRender::HRenderContext render_context, HMaterial material, const RenderObject* ro);
-    void                            ApplyMaterialSampler(dmRender::HRenderContext render_context, HMaterial material, int unit, dmGraphics::HTexture texture);
+    void                            ApplyMaterialSampler(dmRender::HRenderContext render_context, HMaterial material, uint32_t unit, dmGraphics::HTexture texture);
 
     dmGraphics::HProgram            GetMaterialProgram(HMaterial material);
     dmGraphics::HVertexProgram      GetMaterialVertexProgram(HMaterial material);
@@ -327,7 +327,7 @@ namespace dmRender
     bool                            GetMaterialProgramConstantElement(HMaterial material, dmhash_t name_hash, uint32_t element_index, float& out_value);
     void                            SetMaterialProgramConstant(HMaterial material, dmhash_t name_hash, Vectormath::Aos::Vector4 constant);
     int32_t                         GetMaterialConstantLocation(HMaterial material, dmhash_t name_hash);
-    void                            SetMaterialSampler(HMaterial material, dmhash_t name_hash, int16_t unit, dmGraphics::TextureWrap u_wrap, dmGraphics::TextureWrap v_wrap, dmGraphics::TextureFilter min_filter, dmGraphics::TextureFilter mag_filter);
+    void                            SetMaterialSampler(HMaterial material, dmhash_t name_hash, uint32_t unit, dmGraphics::TextureWrap u_wrap, dmGraphics::TextureWrap v_wrap, dmGraphics::TextureFilter min_filter, dmGraphics::TextureFilter mag_filter);
     HRenderContext                  GetMaterialRenderContext(HMaterial material);
 
     uint64_t                        GetMaterialUserData1(HMaterial material);
